+++ math/geometry/distance_test.go
@@ -5,8 +5,8 @@ import (
 )
 
 type args struct {
-	p1 CartesianPoint
-	p2 CartesianPoint
+	p1 EuclideanPoint
+	p2 EuclideanPoint
 }
 
 func TestFindDistanceBetweenTwoPoints(t *testing.T) {
@@ -19,8 +19,8 @@ func TestFindDistanceBetweenTwoPoints(t *testing.T) {
 		{
 			"(0,0) and (2,-2)",
 			args{
-				CartesianPoint{0, 0},
-				CartesianPoint{2, -2},
+				EuclideanPoint{0, 0},
+				EuclideanPoint{2, -2},
 			},
 			2.8284271247461903,
 			false,
@@ -28,8 +28,8 @@ func TestFindDistanceBetweenTwoPoints(t *testing.T) {
 		{
 			"(-20,23) and (-15,68)",
 			args{
-				CartesianPoint{-20, 23},
-				CartesianPoint{-15, 68},
+				EuclideanPoint{-20, 23},
+				EuclideanPoint{-15, 68},
 			},
 			45.27692569068709,
 			false,
@@ -37,8 +37,8 @@ func TestFindDistanceBetweenTwoPoints(t *testing.T) {
 		{
 			"(2,2) and (14,11)",
 			args{
-				CartesianPoint{2, 2},
-				CartesianPoint{14, 11},
+				EuclideanPoint{2, 2},
+				EuclideanPoint{14, 11},
 			},
 			15,
 			false,
@@ -46,7 +46,7 @@ func TestFindDistanceBetweenTwoPoints(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := FindDistanceBetweenTwoPoints(tt.args.p1, tt.args.p2)
+			got, err := EuclideanDistance(tt.args.p1, tt.args.p2)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("EuclideanDistance() error = %v, wantErr %v", err, tt.wantErr)
 				return
@@ -68,8 +68,8 @@ func BenchmarkFindDistanceBetweenTwoPoints(b *testing.B) {
 		{
 			"(0,0) and (2,-2)",
 			args{
-				CartesianPoint{0, 0},
-				CartesianPoint{2, -2},
+				EuclideanPoint{0, 0},
+				EuclideanPoint{2, -2},
 			},
 			2.8284271247461903,
 			false,
@@ -77,8 +77,8 @@ func BenchmarkFindDistanceBetweenTwoPoints(b *testing.B) {
 		{
 			"(-20,23) and (-15,68)",
 			args{
-				CartesianPoint{-20, 23},
-				CartesianPoint{-15, 68},
+				EuclideanPoint{-20, 23},
+				EuclideanPoint{-15, 68},
 			},
 			45.27692569068709,
 			false,
@@ -86,8 +86,8 @@ func BenchmarkFindDistanceBetweenTwoPoints(b *testing.B) {
 		{
 			"(2,2) and (14,11)",
 			args{
-				CartesianPoint{2, 2},
-				CartesianPoint{14, 11},
+				EuclideanPoint{2, 2},
+				EuclideanPoint{14, 11},
 			},
 			15,
 			false,
@@ -97,7 +97,7 @@ func BenchmarkFindDistanceBetweenTwoPoints(b *testing.B) {
 	for _, bm := range benchmarks {
 		b.Run(bm.name, func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				FindDistanceBetweenTwoPoints(bm.args.p1, bm.args.p2)
+				EuclideanDistance(bm.args.p1, bm.args.p2)
 			}
 		})
 	}
